requirements.txt
streamlit>=1.32.0
edge-tts>=6.1.10
moviepy>=1.0.3
pillow>=10.2.0
